+++ src/subapps/subapps.controller.ts
@@ -1,8 +1,15 @@
-import { Controller, Get } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  UseInterceptors,
+  UploadedFile,
+} from '@nestjs/common';
 import { SubappsService } from './subapps.service';
 import { Auth } from 'src/iam/decorators/auth.decorator';
 import { AuthType } from 'src/iam/enums/auth-type.enum';
-
+import { FileInterceptor } from '@nestjs/platform-express';
 @Auth(AuthType.None)
 @Controller('subapps')
 export class SubappsController {
@@ -11,5 +18,13 @@ export class SubappsController {
   @Get()
   getMessage(): string {
     return 'fuck you';
+  }
+
+  // S3 BUCKET UPLOAD
+  // This needs to be moved up to the general subapps controller/service
+  @Post('image-upload')
+  @UseInterceptors(FileInterceptor('image'))
+  async imageUpload(@UploadedFile() file: Express.Multer.File, @Body() body) {
+    return await this.subappsService.imageUpload(file, body);
   }
 }
